Update compare.py
Major overhaul of compare.py. This gets the dissimilarity between the two networks. Since each dissimilarity between the target network and a candidate network is divided by the mean dissimilarity between the target network and 30 (ER) networks. We have two comparisons so that each network gets to be the target network.